+++ server/delphilint-server/src/main/java/au/com/integradev/delphilint/maintenance/FallbackPluginProviderException.java
@@ -1,0 +1,11 @@
+package au.com.integradev.delphilint.maintenance;
+
+public class FallbackPluginProviderException extends Exception {
+  public FallbackPluginProviderException(String message) {
+    super(message);
+  }
+
+  public FallbackPluginProviderException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}